Fixed NAARAD_HOME location in naaradboot.github.sh. Added Apps/naaradimport.py.  Seems an important file to miss! Adding Boot/setHostName file.
Attempt to fix the RT bug that gets the server in a infinite loop,
possibly due to a corrupted over-the-air packet.  This should not lead
to a problem (and it did not till about 3 years ago when the code was
ported to python3 -- which had important changes in how strings are coded/decoded
and handled):

Checking if the output of uno.readline() in NaaradTopic2::run() reports
a COM port timeout.  If it does, report and set the output to blank
JSON ('{}').