Capability to self reboot.
Self-reboot is now initiated if uno.readline() returns a blank string
which is supposed to indicate COM port read-timeout.  When this is
detected, send the shutdown command twice on localpost socket 1234.
With this the sever shutsdown and initServer() call in naarad.py
exits, allowing a reboot sequence to begin.

In naarad.py __main__, rapid reboots (reboots within 1 min of each
other) is limited to 5 tries.

Adding the ArduinoPortDiscovery.py to programmatically discover
the COM port on which Arduino appears.  This functions is not
yet in use.